Working example
This is a really simple example use of the iShip API using javascript as the client and simply displays the price that the shipment was quoted at. There is more information that you can show but i felt for a start returning a price was plenty.

this example code for a javascript connection into the iShip API uses code from
https://github.com/douglascrockford/JSON-js/